refactor: Home component 분리
