+++ src/pages/Home/SlideCardSection/styled.ts
@@ -1,0 +1,2 @@
+import styled from 'styled-components';
+import { motion } from 'framer-motion';
