+++ src/pages/Home/WelcomePharaseText/WelcomePhraseText.tsx
